Rename css-variables to match Architect
