+++ src/renderer/components/Transitions/SlideDown.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { CSSTransition } from 'react-transition-group';
 
-import { getCSSNumber } from '../../utils/css-variables';
+import { getCSSNumber } from '../../utils/CSSVariables';
 
 // Assume user is only overriding custom props at root
 const duration = getCSSNumber('--animation-duration-fast-ms');
